refactor: extract resolve_app_param_name, remove FlatFileArgs, make app_param arg consistent for all modes
