+++ crates/bevy_auto_plugin_proc_macros/src/docs/flat_file/auto_plugin.md
@@ -1,7 +1,7 @@
 The main attribute for flat file mode that processes all auto attributes in the file.
 
 # Parameters
-- `app=identifier` - Optional. Specifies the app variable name that code will be injected into. (defaults to app)
+- `app_param=identifier` - Optional. Specifies the app variable name that code will be injected into. (defaults to app)
 
 # Example
 ```rust
@@ -15,7 +15,7 @@ use bevy_auto_plugin::modes::flat_file::prelude::*;
 // struct MyComponent;
 
 // ^ auto macro attributes must be declared above #[auto_plugin]
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {
     // Code generated by the macro is injected here.
     // For example:
