+++ crates/bevy_auto_plugin_proc_macros/src/docs/flat_file/auto_register_type.md
@@ -14,7 +14,7 @@ use bevy_auto_plugin::modes::flat_file::prelude::*;
 #[reflect(Component)]
 struct FooComponent;
 
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {
     /* generated code */
     // app.register_type::<FooComponent>();
@@ -32,7 +32,7 @@ use bevy_auto_plugin::modes::flat_file::prelude::*;
 #[reflect(Component)]
 struct FooComponentWithGeneric<T>(T);
 
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {
     /* generated code */
     // app.register_type::<FooComponentWithGeneric<bool>>();
