+++ crates/bevy_auto_plugin_shared/src/__private/modes/flat_file/mod.rs
@@ -1,3 +1,2 @@
-pub mod attribute;
 pub mod file_state;
 pub mod inner;
