+++ tests/flat_file/auto_add_event.rs
@@ -6,7 +6,7 @@ use bevy_ecs::prelude::*;
 #[derive(Event)]
 struct Test;
 
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {}
 
 fn app() -> App {
