+++ tests/flat_file/auto_add_systems.rs
@@ -10,7 +10,7 @@ fn foo_system(mut test: ResMut<Test>) {
     test.0 += 1;
 }
 
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {
     app.init_resource::<Test>();
 }
