+++ tests/flat_file/auto_add_systems_with_generics.rs
@@ -38,7 +38,7 @@ where
     test.0 += T::ONE;
 }
 
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {
     app.init_resource::<Test<u32>>();
     app.init_resource::<Test<i32>>();
