+++ tests/flat_file/auto_init_resource.rs
@@ -6,7 +6,7 @@ use bevy_ecs::prelude::*;
 #[derive(Resource, Default)]
 struct Test;
 
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {}
 
 fn app() -> App {
