+++ tests/flat_file/auto_name.rs
@@ -7,7 +7,7 @@ use bevy_ecs::prelude::*;
 #[auto_name]
 pub struct Test;
 
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {}
 
 fn app() -> App {
