+++ tests/flat_file/auto_name_with_generics.rs
@@ -7,7 +7,7 @@ use bevy_ecs::prelude::*;
 #[auto_name(generics(bool))]
 pub struct Test<T>(T);
 
-#[auto_plugin(app=app)]
+#[auto_plugin(app_param=app)]
 fn plugin(app: &mut App) {}
 
 fn app() -> App {
