+++ tests/flat_file/auto_plugin_param.rs
@@ -8,7 +8,7 @@ use internal_test_util::type_id_of;
 #[derive(Reflect)]
 struct Test;
 
-#[auto_plugin(app=my_app)]
+#[auto_plugin(app_param=my_app)]
 fn plugin(my_app: &mut App) {}
 
 fn app() -> App {
