+++ tests/flat_file/ui/auto_plugin_already_registered.rs
@@ -1,6 +1,6 @@
 use bevy_auto_plugin::modes::flat_file::prelude::*;
 
-#[auto_plugin(app=_app)]
+#[auto_plugin(app_param=_app)]
 fn plugin(_app: &mut bevy_app::App) {}
 
 #[auto_register_type]
