+++ tests/flat_file/ui/nightly/invalid_generics.rs
@@ -3,7 +3,7 @@ use bevy_auto_plugin::modes::flat_file::prelude::*;
 #[auto_register_type(generics(u32))]
 struct Test<T1, T2>(T1, T2);
 
-#[auto_plugin(app=_app)]
+#[auto_plugin(app_param=_app)]
 fn plugin(_app: &mut bevy_app::App) {}
 
 // dummy main
