+++ tests/flat_file/ui/stable/auto_plugin_app_not_mutable.rs
@@ -1,6 +1,6 @@
 use bevy_auto_plugin::modes::flat_file::prelude::*;
 
-#[auto_plugin(app=_app)]
+#[auto_plugin(app_param=_app)]
 fn plugin(_app: &bevy_app::App) {}
 
 // dummy main
